src/requirements.txt
annotated-types==0.7.0
anthropic==0.52.0
anyio==4.9.0
async-timeout==4.0.3
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
distro==1.9.0
exceptiongroup==1.3.0
fpdf==1.7.2
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-anthropic==0.3.13
langchain-core==0.3.61
langchain-text-splitters==0.3.8
langgraph==0.4.5
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.0
langgraph-sdk==0.1.70
langsmith==0.3.42
orjson==3.10.18
ormsgpack==1.9.1
packaging==24.2
pillow==11.2.1
pydantic==2.11.4
pydantic-core==2.33.2
python-dotenv==1.1.0
pyyaml==6.0.2
reportlab==4.4.1
requests==2.32.3
requests-toolbelt==1.0.0
ruff==0.11.10
sniffio==1.3.1
sqlalchemy==2.0.41
tenacity==9.1.2
typing-extensions==4.13.2
typing-inspection==0.4.1
urllib3==2.4.0
xxhash==3.5.0
zstandard==0.23.0
